Refactor ContractCallType * Replace ContractCallType to qrc20.rs * Refactor ContractCallType::from_script_pubkey()
